Convert to pyproject.toml package + bump some deps; automate publishing
