+++ NEWS
@@ -1,0 +1,5 @@
+## Changelog
+
+### 0.1.0
+
+First release of `sopel-deepl` with basic translate-to-English functionality.
