--- setup.cfg
@@ -1,34 +1,0 @@
-[metadata]
-name = sopel-deepl
-version = 0.1.0
-description = DeepL translation plugin for Sopel IRC bots.
-author = dgw
-author_email = dgw@technobabbl.es
-url = https://github.com/dgw/sopel-deepl
-license = Eiffel Forum License, version 2
-classifiers =
-    Intended Audience :: Developers
-    Intended Audience :: System Administrators
-    License :: Eiffel Forum License (EFL)
-    License :: OSI Approved :: Eiffel Forum License
-    Programming Language :: Python :: 3 :: Only
-    Programming Language :: Python :: 3.6
-    Programming Language :: Python :: 3.7
-    Programming Language :: Python :: 3.8
-    Programming Language :: Python :: 3.9
-    Programming Language :: Python :: 3.10
-    Topic :: Communications :: Chat :: Internet Relay Chat
-
-[options]
-packages = find:
-zip_safe = false
-include_package_data = true
-install_requires =
-    sopel>=7.1,<9
-    deepl_api~=0.2.2
-# deepl_api requires py3.7, otherwise I'd do 3.6+
-python_requires = >=3.7,<4
-
-[options.entry_points]
-sopel.plugins =
-    deepl = sopel_deepl
